Finished cognito flow using localStorage to sync multiple tabs

+++ src/index.js
@@ -11,26 +11,8 @@ import "antd/dist/antd.css";
 import "./index.css";
 
 
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
 if (window.screen.width < 1000) {
-  message.info("For a better user experience, please consider visiting this site on a computer :)", 8);
+  message.info("For a better user experience, please consider visiting this site on a computer :)", 5);
 }
 
 
@@ -39,51 +21,45 @@ const initialLife = {
   lifespan: 0
 };
 
-// const testLife = {
-  //   DOB: dayjs('1990-04-16'),
-  //   lifespan: 100
-  // };
+const testLife = {
+    DOB: dayjs('1995-04-16'),
+    lifespan: 30
+  };
   
 function Index() {
-  const [basicInfo, setBasicInfo] = useState({});
-  const [life, setLife] = useState(initialLife);
+  const [basicInfo, setBasicInfo] = useState({ name: ' ', gender: ' '});
+  const [life, setLife] = useState(testLife);
   const [epochs, setEpochs] = useState([]);
   const [user, setUser] = useState(undefined);
-  const FB = useRef(undefined);
 
-
-  document.addEventListener('FBObjectReady', () => FB.current = window.FB);
   window.addEventListener('load', initUser);
-    
-
+  window.addEventListener('storage', () => {
+    const loggedIn = window.localStorage.getItem('loggedIn');
+    if (!loggedIn && user) {
+      auth.signOut();
+    }
+  });
 
   function initUser() {
     let currentUser;
+    const loggedIn = window.localStorage.getItem('loggedIn');
     const hash = window.location.hash;
-    if (hash) {
+    if (hash && hash !== '#') {
+      window.location = '#';
       currentUser = new User(hash);
-  
-    } else {
-      let idToken = window.localStorage.getItem('idToken');
-      let accessToken = window.localStorage.getItem('accessToken');
-      if (idToken && accessToken) {
-        currentUser = new User('', idToken, accessToken);
-      }
-    }
-
-    if (!currentUser) {
-      window.localStorage.clear();
+    } else { 
+      if(!loggedIn) return;
+      auth.getSession();
       return;
     }
 
-    try {
-      setUser(currentUser);
-      window.localStorage.setItem('idToken', currentUser.idToken);
-      window.localStorage.setItem('accessToken', currentUser.accessToken);
-
-    } catch(err) {
-      if (!err) alert('login success');
+    if (!currentUser) {
+      if (loggedIn) window.localStorage.clear();
+      return;
     }
+
+    setUser(currentUser);
+    if (!loggedIn) window.localStorage.setItem('loggedIn', '1');
   }
 
   const importUserData = () => {
@@ -92,28 +68,29 @@ function Index() {
       if (!saved) return;
 
       setBasicInfo({
-        name: saved.name || undefined,
-        gender: saved.gender || undefined
+        name: saved.name || ' ',
+        gender: saved.gender || ' ' 
       });
 
       setLife({
-        DOB: dayjs(saved.DOB || ''),
+        DOB: dayjs(saved.DOB),
         lifespan: saved.lifespan || 0
       });
       
-      setEpochs(saved.epochs? 
-        [...saved.epochs.map( item => {
+      setEpochs(saved.epochs.map( item => {
           return {
             uuid: item.uuid,
             title: item.title,
-            note: item.note || '',
+            note: item.note || ' ',
             color: item.color,
             start: dayjs(item.start),
             end: dayjs(item.end)
           };
-        })] : []
+        })
       );
-    })
+    }).catch( err => {
+      console.log(`Error: ${err}`);
+    });
   }
 
   const login = () => {
@@ -129,108 +106,30 @@ function Index() {
     user.deleteUser();
     logout();
   }
-
-  useEffect( () => {
-    if (!user) return;
-
-    importUserData();
-
-    // if(FB.current && user.fbId) {
-    //   FB.current.api(
-    //     '/'+user.fbId,
-    //     'GET',
-    //     {"fields":"first_name,middle_name,last_name"},
-    //     res => {
-    //       alert(res.toString());
-    //       // loginNotice(first_name, last_name);
-    //       // user.update({
-    //       //   "Fb_first_name": first_name,
-    //       //   "Fb_middle_name": middle_name,
-    //       //   "Fb_last_name": last_name
-    //       // });
-    //     }
-      // );
-    // }
-  }, [user]);
   
-
-  // function handleFbLogin(loginButtonClick) {
-  //   return (res) => {
-  //     if (res.status === 'connected') {
-  //       const queryUserFb = new Promise(resolve => {
-  //         FB.api(
-  //           '/'+res.authResponse.userID,
-  //           'GET',
-  //           {"fields":"id,first_name,middle_name,last_name"},
-  //           res => resolve(res)
-  //         );
-  //       });
-        
-  //       const queryUserDb = getUser(res.authResponse.userID);
-        
-  //       Promise.all([queryUserFb, queryUserDb]).then( values => {
-  //         const fbUser = values[0];
-  //         const dbUser = values[1].data.Item;
-          
-  //         if (!fbUser.id) {
-  //           return setTimeout(() => FB.login(handleFbLogin(true)), 0);
-  //         }
-                    
-  //         if (dbUser && (dbUser.loginStatus || loginButtonClick) ) {
-  //           updateUser({
-  //             Key: { "UserID": dbUser.UserID },
-  //             UpdateExpression: "set #a = :a, #b = :b",
-  //             ExpressionAttributeNames: {
-  //               "#a": "FbInfo",
-  //               "#b": "loginStatus"
-  //             },
-  //             ExpressionAttributeValues: {
-  //               ":a": {
-  //               "firstName": fbUser.first_name,
-  //               "middleName": fbUser.middle_name,
-  //               "lastName": fbUser.last_name   
-  //               },
-  //               ":b": true
-  //             }
-  //           });
-
-  //         } else if (!dbUser && loginButtonClick) {
-  //           createUser({
-  //             UserID: fbUser.id,
-  //             FbInfo: {
-  //               "firstName": fbUser.first_name,
-  //               "middleName": fbUser.middle_name,
-  //               "lastName": fbUser.last_name
-  //             },
-  //             loginStatus: true
-  //           }); 
-            
-  //         } else {
-  //           setUser(undefined);
-  //           return;
-  //         }
-          
-  //         setUser(fbUser);
-  //         loadData(dbUser);
-  //         loginNotice(fbUser);
-  //       });
-  //     } else {
-  //       setUser(undefined);
-  //     }
-  //   };
-  // }
-
-  // function handleLoginBtnClick() {
-  //   FB.login(handleFbLogin(true));
-  // }
-  
-  function loginNotice(firstName, lastName) {
+  const save = () => {
+    user.save({
+        UserID: user.id,
+        name: basicInfo.name,
+        gender: basicInfo.gender,
+        DOB: life.DOB,
+        lifespan: life.lifespan,
+        epochs: epochs
+      }).then( res => {
+        message.success('Saved!', 2);
+      }).catch( err => {
+        console.log(`Error: ${err}`);
+      })
+    }
+    
+  const loginNotice = (fbName) => {
     message.success(
       <span>
-        {`Logged in as ${firstName} ${lastName}`}
+        {`Logged in as ${fbName}`}
         <br/>
-        <img 
-          src={`http://graph.facebook.com/${user.fbId}/picture?type=large`}
+        <img
+          style={{paddingTop: '1rem', paddingBottom: '0.7rem'}}
+          src={`http://graph.facebook.com/${user.fbId}/picture?type=normal`}
           alt=''
         />
       </span>, 
@@ -238,30 +137,17 @@ function Index() {
     );
   }
 
-  const save = () => {
-    if (life.lifespan === 0) {
-      message.info('Nothing new to save.', 3);
-      return;
-    }
+    useEffect( () => {
+      if (!user) return;
+  
+      importUserData();
+      setTimeout(()=> loginNotice(user.fbName), 600);
 
-    user.update({
-      "name": basicInfo.name,
-      "gender": basicInfo.gender,
-      "DOB": life.DOB.toString(),
-      "lifespan": life.lifespan,
-      "epochs": epochs
-    }).then( (_, err) => {
-      if (err) {
-        alert(err);
-      } else {
-        message.success('Saved!', 3);
-      }
-    });
-  }
+    }, [user]);
+    
 
-
-  return (
-    <div className='MainContainer'>   
+    return (
+      <div className='MainContainer'>   
       <div className='App'>
         <UserInfo setBasicInfo={setBasicInfo} setLife={setLife}/>
         <Calendar life={life} epochs={epochs} setEpochs={setEpochs}/>
